Transition to CGFloat
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@3196 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Terminal/TerminalView.m
@@ -2379,7 +2379,7 @@ improve? */
 	ignore_resize=ignore;
 }
 
--(void) setBorder: (float)x : (float)y
+-(void) setBorder: (CGFloat)x : (CGFloat)y
 {
 	border_x=x;
 	border_y=y;
